test: split phpunit and pest tests, add arch tests
