+++ Makefile
@@ -5,6 +5,10 @@
 # https://docs.docker.com/compose/environment-variables/envvars/
 export DOCKER_BUILDKIT ?= 1
 
+# User ID and Group ID to use inside docker containers
+HOST_UID ?= $(shell id -u)
+HOST_GID ?= $(shell id -g)
+
 # Docker binary to use, when executing docker tasks
 DOCKER ?= docker
 
@@ -53,7 +57,9 @@ NPM_RUNNER ?= pnpm
 
 EXPORT_VARS = '\
 	$${COMPOSE_PROJECT_NAME} \
-	$${COMPOSER_AUTH}'
+	$${COMPOSER_AUTH} \
+	$${HOST_UID} \
+	$${HOST_GID}'
 
 #
 # Self documenting Makefile code
@@ -92,9 +98,9 @@ help: ## Show this menu
 	@echo
 	@echo '    📑 Logs are stored in      $(MAKE_LOGFILE)'
 	@echo
-	@echo '    📦 Package                 active-record (github.com/cycle/active-record)'
-	@echo '    🤠 Makefile Author         Andrij Orlenko (github.com/lotyp)'
-	@echo '    🏢 ${YELLOW}Org                     cycle (github.com/cycle)${RST}'
+	@echo '    📦 Package                 active-record (https://github.com/cycle/active-record)'
+	@echo '    🤠 Makefile Author         Andrij Orlenko (https://github.com/lotyp)'
+	@echo '    🏢 ${YELLOW}Org                     cycle (https://github.com/cycle)${RST}'
 	@echo
 .PHONY: help
 
@@ -123,7 +129,7 @@ else
 endif
 .PHONY: env
 
-prepare:
+prepare: ## Prepare project for development
 	mkdir -p .build/php-cs-fixer
 .PHONY: prepare
 
@@ -251,7 +257,7 @@ lint-deps: ## Runs composer-require-checker – checks for dependencies that are
 .PHONY: lint-deps
 
 lint-composer: ## Normalize composer.json and composer.lock files
-	$(APP_COMPOSER) normalize
+	$(APP_RUNNER) .phive/composer-normalize normalize
 .PHONY: lint-composer
 
 lint-audit: ## Runs security checks for composer dependencies
@@ -273,9 +279,16 @@ infect-ci: ## Runs infection – mutation testing framework with github output (
 	$(APP_COMPOSER) infect:ci
 .PHONY: lint-infect-ci
 
+test-all: test test-arch test-pgsql test-mysql test-sqlite test-sqlserver ## Run all test suites
+.PHONY: test-all
+
 test: ## Run project php-unit and pest tests
 	$(APP_COMPOSER) test
 .PHONY: test
+
+test-arch: ## Run project pest tests with architecture checks
+	$(APP_COMPOSER) test:arch
+.PHONY: test-arch
 
 test-pgsql: ## Run project php-unit and pest tests over pgsql database
 	$(APP_COMPOSER) test:pgsql
@@ -300,6 +313,6 @@ test-cc: ## Run project php-unit and pest tests in coverage mode and build repor
 #
 # Release
 # ------------------------------------------------------------------------------------
-commit:
+commit: ## Run commitizen to create commit message
 	czg commit --config="./.github/.cz.config.js"
 .PHONY: commit
