+++ phpstan-baseline.neon
@@ -3,4 +3,4 @@ parameters:
 		-
 			message: "#^Call to an undefined method Cycle\\\\Database\\\\Driver\\\\DriverInterface\\:\\:setLogger\\(\\)\\.$#"
 			count: 1
-			path: tests/src/DatabaseTestCase.php
+			path: tests/src/Functional/DatabaseTestCase.php
