+++ psalm.xml
@@ -17,7 +17,10 @@
 
     <projectFiles>
         <directory name="src/"/>
-        <directory name="tests/"/>
+        <directory name="tests/src/"/>
+        <ignoreFiles>
+            <directory name="tests/src/Arch/"/>
+        </ignoreFiles>
         <file name=".php-cs-fixer.dist.php"/>
         <file name="rector.php"/>
         <ignoreFiles>
