+++ tests/src/Functional/FacadeTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Cycle\Tests;
+namespace Cycle\Tests\Functional;
 
 use Cycle\ActiveRecord\Exception\ConfigurationException;
 use Cycle\ActiveRecord\Facade;
