+++ tests/src/Functional/Loggable.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Cycle\Tests;
+namespace Cycle\Tests\Functional;
 
 use Cycle\Database\Driver\DriverInterface;
