+++ tests/src/Functional/Query/ActiveQueryTest.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Cycle\Tests\Query;
+namespace Cycle\Tests\Functional\Query;
 
 use Cycle\ActiveRecord\Facade;
 use Cycle\App\Entity\User;
 use Cycle\Database\Database;
 use Cycle\Database\DatabaseManager;
-use Cycle\Tests\DatabaseTestCase;
+use Cycle\Tests\Functional\DatabaseTestCase;
 use PHPUnit\Framework\Attributes\Test;
 
 final class ActiveQueryTest extends DatabaseTestCase
