+++ tests/src/Functional/TestLogger.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Cycle\Tests;
+namespace Cycle\Tests\Functional;
 
 use Psr\Log\LoggerInterface;
 use Psr\Log\LoggerTrait;
