--- tests/src/Pest.php
@@ -1,7 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Cycle\Tests;
-
-uses(TestCase::class)->in(__DIR__);
